Remove unused dependency (#19)
Due to a cargo bug, this was causing issues with adding rusqlite to ZQ2. See https://github.com/launchbadge/sqlx/issues/3211 - Cargo would detect a spurious version conflict in the sqlite version to bundle.
However sqlx appears completely unused in zqutils so removing the dependency entirely is an easy fix.